requirements.txt
flask
flask_session
gunicorn
mysql.connector